requirements.txt
streamlit>=1.40
openai>=1.50
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
python-dotenv>=1.0
